+++ src/Generators/Tests/Laravel/Listener/ListenerTestGenerator.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpUnitGen\Core\Generators\Tests\Laravel\Listener;
+
+use PhpUnitGen\Core\Generators\Tests\Laravel\Job\JobTestGenerator;
+
+/**
+ * Class ListenerTestGenerator.
+ *
+ * @author  Paul Thébaud <paul.thebaud29@gmail.com>
+ * @author  Killian Hascoët <killianh@live.fr>
+ * @license MIT
+ */
+class ListenerTestGenerator extends JobTestGenerator
+{
+}