+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 **Added**
 
 - Add a test generator for Laravel Controller.
+- Add a test generator for Laravel Job.
+- Add a test generator for Laravel Listener.
 
 **Changed**